convert Eigen::Vector3d to Eigen::Vector3f in Class Point

+++ src/model/Point.h
@@ -22,5 +22,5 @@ public:
 		position.z() = z;
 	}
 
-	Eigen::Vector3d position;
+	Eigen::Vector3f position;
 };
